Add files via upload
"I created a website that allows users to design their own custom shoes using a range of colors, materials, and design options. The website was built using HTML, CSS, JavaScript, and Bootstrap, showcasing my skills in web development and design. With its intuitive interface and unique customization features, the website has received positive feedback from users and has become a go-to destination for shoe enthusiasts looking for one-of-a-kind designs."